Moved constants to constants file && removed unused files
