+++ src/constants/constants.js
@@ -2,14 +2,19 @@ const CONSTANTS = {};
 
 
 CONSTANTS.LABELS = {
-  BUTTON: 'Button'
+  BUTTON: 'Button',
+  OK: 'Ok',
+  CANCEL: 'Cancel'
 };
 
 CONSTANTS.PLACEHOLDER = {
   CARD: 'This is card placeholder text',
+  FIELD_ID: 'fieldId',
   INPUT: 'Enter text',
+  INPUT_ID: 'inputId',
   HEADER: 'Heading',
   POPUP: 'This is a sample popup',
+  POPUP_TITLE: 'Are you sure?',
   TEXT: 'Text',
   TEXT_PRIMARY: 'Text Primary'
 };
