--- src/stories/8-ReceiveLoader.stories.js.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import { ReceiveLoader } from '../components/core';
-
-
-export default {
-  title: 'ReceiveLoader'
-};
-
-export const receiveLoader = () => <ReceiveLoader />
